requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
python-dotenv
pypdf
sentence-transformers